requirements.txt
absl-py==2.2.2
alembic==1.15.2
catboost==1.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
idna==3.4
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lightgbm==4.6.0
lightning-utilities==0.14.3
Mako==1.3.9
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
narwhals==1.34.1
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pprintpp==0.4.0
protobuf==6.30.2
pyarrow==19.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-frame==0.2.5
pytz==2025.2
PyYAML==6.0.2
requests==2.28.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.40
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0

tqdm==4.67.1
triton==2.2.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==1.26.13
Werkzeug==3.1.3
xgboost==2.1.4
zipp==3.21.0
